Introduce a collections crate w/ deterministic hashmap, hashset in tests
